requirements.txt
Pillow==7.2.0
django==2.1
django-ckeditor==5.9.0
django-colorfield==0.3.2
python-decouple==3.3
xhtml2pdf==0.2.5
pylint==2.5.3